ticket_backend/requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
databases==0.5.3
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
jose==1.0.0
ormar==0.10.22
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
python-ulid==1.0.3
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
